+++ examples/json-extension/server/server.py
@@ -17,6 +17,7 @@
 import asyncio
 import json
 import time
+import uuid
 from json import JSONDecodeError
 
 from pygls.features import (COMPLETION, TEXT_DOCUMENT_DID_CHANGE,
@@ -26,7 +27,9 @@ from pygls.types import (CompletionItem, CompletionList, CompletionParams,
                          ConfigurationItem, ConfigurationParams, Diagnostic,
                          DidChangeTextDocumentParams,
                          DidCloseTextDocumentParams, DidOpenTextDocumentParams,
-                         Position, Range)
+                         MessageType, Position, Range, Registration,
+                         RegistrationParams, Unregistration,
+                         UnregistrationParams)
 
 COUNT_DOWN_START_IN_SECONDS = 10
 COUNT_DOWN_SLEEP_IN_SECONDS = 1
@@ -35,9 +38,11 @@ COUNT_DOWN_SLEEP_IN_SECONDS = 1
 class JsonLanguageServer(LanguageServer):
     CMD_COUNT_DOWN_BLOCKING = 'countDownBlocking'
     CMD_COUNT_DOWN_NON_BLOCKING = 'countDownNonBlocking'
+    CMD_REGISTER_COMPLETIONS = 'registerCompletions'
     CMD_SHOW_CONFIGURATION_ASYNC = 'showConfigurationAsync'
     CMD_SHOW_CONFIGURATION_CALLBACK = 'showConfigurationCallback'
     CMD_SHOW_CONFIGURATION_THREAD = 'showConfigurationThread'
+    CMD_UNREGISTER_COMPLETIONS = 'unregisterCompletions'
 
     CONFIGURATION_SECTION = 'jsonServer'
 
@@ -139,6 +144,19 @@ async def did_open(ls, params: DidOpenTextDocumentParams):
     _validate(ls, params)
 
 
+@json_server.command(JsonLanguageServer.CMD_REGISTER_COMPLETIONS)
+async def register_completions(ls: JsonLanguageServer, *args):
+    """Register completions method on the client."""
+    params = RegistrationParams([Registration(str(uuid.uuid4()), COMPLETION,
+                                              {"triggerCharacters": "[':']"})])
+    response = await ls.register_capability_async(params)
+    if response is None:
+        ls.show_message('Successfully registered completions method')
+    else:
+        ls.show_message('Error happened during completions registration.',
+                        MessageType.Error)
+
+
 @json_server.command(JsonLanguageServer.CMD_SHOW_CONFIGURATION_ASYNC)
 async def show_configuration_async(ls: JsonLanguageServer, *args):
     """Gets exampleConfiguration from the client settings using coroutines."""
@@ -194,3 +212,15 @@ def show_configuration_thread(ls: JsonLanguageServer, *args):
 
     except Exception as e:
         ls.show_message_log('Error ocurred: {}'.format(e))
+
+
+@json_server.command(JsonLanguageServer.CMD_UNREGISTER_COMPLETIONS)
+async def unregister_completions(ls: JsonLanguageServer, *args):
+    """Unregister completions method on the client."""
+    params = UnregistrationParams([Unregistration(str(uuid.uuid4()), COMPLETION)])
+    response = await ls.unregister_capability_async(params)
+    if response is None:
+        ls.show_message('Successfully unregistered completions method')
+    else:
+        ls.show_message('Error happened during completions unregistration.',
+                        MessageType.Error)